refactor: restructure repository #56
