--- ISSUES.md
@@ -1,20 +1,0 @@
-# Issues
-
-- [x] #20: Scaffold `sha` CLI project
-- [x] #22: Implement `sha new` command
-- [x] #24: Implement `sha add` command
-- [x] #26: Implement `sha version` command
-- [x] #28: Implement `sha env` command
-- [x] #30: Implement `sha run`, `sha build`, `sha test` wrappers
-- [x] #32: Implement `sha flash` and `sha deploy`
-- [x] #34: Implement Code-First API Generation
-- [x] #35: Implement Modular CI/CD Standards
-- [x] #36: Implement Observability & Logging (The Pulse)
-- [x] #37: Implement ML & Research Rigor Features
-- [x] #38: Optimize for rustup docs
-- [x] #44: Cross-platform one-liner install script
-- [x] #45: Auto-installation of workspace dependencies
-- [ ] #48: Add UI & Table dependencies
-- [ ] #49: Visual Overhaul: No Emojis & Color-coding
-- [ ] #50: Grid Layout Integration
-- [ ] #51: Implement `sha issue` subcommand
