+++ cli/CHANGELOG.md
@@ -1,0 +1,50 @@
+# [1.4.0](https://github.com/shawal-mbalire/shastack/compare/v1.3.0...v1.4.0) (2026-04-30)
+
+
+### Features
+
+* enhance CLI with auto-init and populate modules with boilerplate [#18](https://github.com/shawal-mbalire/shastack/issues/18) ([ed8ac1f](https://github.com/shawal-mbalire/shastack/commit/ed8ac1faa58a644b7755ecaddc56bb4b9106e056))
+
+# [1.3.0](https://github.com/shawal-mbalire/shastack/compare/v1.2.0...v1.3.0) (2026-04-30)
+
+
+### Features
+
+* **cli:** rewrite shastack CLI in Rust [#17](https://github.com/shawal-mbalire/shastack/issues/17) ([c8d9576](https://github.com/shawal-mbalire/shastack/commit/c8d95763fd78dcb119a188588952f1930b45e8c2))
+* complete rust pivot with one-liner install [#16](https://github.com/shawal-mbalire/shastack/issues/16) ([93b5ff2](https://github.com/shawal-mbalire/shastack/commit/93b5ff220ef5678e3474cfd754c252ffd0b8a978))
+
+# [1.2.0](https://github.com/shawal-mbalire/shastack/compare/v1.1.0...v1.2.0) (2026-04-30)
+
+
+### Features
+
+* **cli:** register cli as a formal module [#15](https://github.com/shawal-mbalire/shastack/issues/15) ([883e434](https://github.com/shawal-mbalire/shastack/commit/883e43460ac6329118557569bdc55472c970b721))
+
+# [1.1.0](https://github.com/shawal-mbalire/shastack/compare/v1.0.0...v1.1.0) (2026-04-30)
+
+
+### Features
+
+* **cli:** add global --dry-run option [#13](https://github.com/shawal-mbalire/shastack/issues/13) ([22af6e2](https://github.com/shawal-mbalire/shastack/commit/22af6e2e72a36fdd4781da2b9fa30c5256a061e1))
+
+# 1.0.0 (2026-04-30)
+
+
+### Features
+
+* add initial specifications for shastack framework ([2b10767](https://github.com/shawal-mbalire/shastack/commit/2b107674b195b60b44a1e1c97f0f715d4e24b4f1))
+* **cli:** add audit command [#3](https://github.com/shawal-mbalire/shastack/issues/3) ([fa6ec6d](https://github.com/shawal-mbalire/shastack/commit/fa6ec6d3e8fc6fae08d656a1bd0d5b1a37c66d58))
+* **cli:** add benchmarking command [#2](https://github.com/shawal-mbalire/shastack/issues/2) ([7580641](https://github.com/shawal-mbalire/shastack/commit/75806418dfa676f8ccddca4802a1817b99e21317))
+* **cli:** add docs command [#4](https://github.com/shawal-mbalire/shastack/issues/4) ([481d4da](https://github.com/shawal-mbalire/shastack/commit/481d4da9e2f99d736b2baabdeb4794a96295e459))
+* **cli:** add env management [#04](https://github.com/shawal-mbalire/shastack/issues/04) ([b99e36c](https://github.com/shawal-mbalire/shastack/commit/b99e36cceb539e035a8c152ce2ddf3172ca7711f))
+* **cli:** add event bus command [#1](https://github.com/shawal-mbalire/shastack/issues/1) ([4afde41](https://github.com/shawal-mbalire/shastack/commit/4afde414d8cf8a3c9df89924ea3bdcaf436a11d0))
+* **cli:** add release command [#5](https://github.com/shawal-mbalire/shastack/issues/5) ([eb80333](https://github.com/shawal-mbalire/shastack/commit/eb8033311b54f0b48a9275616d5e86d3d914ba56))
+* **cli:** implement core sha new and sha run [#03](https://github.com/shawal-mbalire/shastack/issues/03) ([a3b5c39](https://github.com/shawal-mbalire/shastack/commit/a3b5c394c846c5c9050233ae99a541e9f1998c0c))
+* **hardware:** scaffold hardware module [#08](https://github.com/shawal-mbalire/shastack/issues/08) ([ea481a6](https://github.com/shawal-mbalire/shastack/commit/ea481a67c228b6ead6e17e9843adc913b3bf5a8b))
+* implement automated semver ci with conventional commits [#11](https://github.com/shawal-mbalire/shastack/issues/11) ([ed8854e](https://github.com/shawal-mbalire/shastack/commit/ed8854e695811203d5c8e9b95e6e08bef24c9698))
+* **infra:** scaffold infra module [#10](https://github.com/shawal-mbalire/shastack/issues/10) ([82259be](https://github.com/shawal-mbalire/shastack/commit/82259bec88f3ee86f20dd52baf2a4c40680aeb32))
+* **meta:** initial scaffolding and roadmap [#01](https://github.com/shawal-mbalire/shastack/issues/01) ([06a512f](https://github.com/shawal-mbalire/shastack/commit/06a512f30cf5e59e2775ea6d8e4fd8e8610764f8))
+* **ml:** scaffold ml module [#06](https://github.com/shawal-mbalire/shastack/issues/06) ([3a822b1](https://github.com/shawal-mbalire/shastack/commit/3a822b16f104e4a1d3c263da3d2a1c298b4b9fc3))
+* **mobile:** scaffold mobile module [#09](https://github.com/shawal-mbalire/shastack/issues/09) ([3330960](https://github.com/shawal-mbalire/shastack/commit/3330960a8a5c87978488f697a2e0ced2f9623cbb))
+* **research:** scaffold research module [#07](https://github.com/shawal-mbalire/shastack/issues/07) ([7c3e7a0](https://github.com/shawal-mbalire/shastack/commit/7c3e7a0e788c8b36f996921e19a0a7f03cd4b411))
+* **web:** scaffold web module [#05](https://github.com/shawal-mbalire/shastack/issues/05) ([28ae909](https://github.com/shawal-mbalire/shastack/commit/28ae90912def59ddc6a5235f1195aba4358b8c79))
