+++ cli/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "sha"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = "1.0.102"
+chrono = "0.4.44"
+clap = { version = "4.6.1", features = ["derive"] }
+colored = "3.1.1"
+comfy-table = "7.2.2"
+inquire = "0.9.4"
+semver = "1.0.28"
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.149"
+
+[dev-dependencies]
+tempfile = "3.27.0"
