+++ cli/README.md
@@ -1,0 +1,21 @@
+# shastack
+
+The Unified Universal Stack Specification CLI.
+
+## Installation
+
+### Unix (macOS / Linux)
+```bash
+curl -sSfL https://raw.githubusercontent.com/shawal-mbalire/shastack/main/scripts/install.sh | bash
+```
+
+### Windows
+```powershell
+powershell -c "irm https://raw.githubusercontent.com/shawal-mbalire/shastack/main/scripts/install.ps1 | iex"
+```
+
+## Usage
+
+```bash
+sha --help
+```
