+++ docs/README.md
@@ -1,0 +1,2 @@
+# shastack Documentation
+Documentation for the Unified Universal Stack Specification.
