+++ landing-page/README.md
@@ -1,0 +1,2 @@
+# shastack Landing Page
+The web presence for the shastack ecosystem.
